Allow DbParams.from_environment to use any name
This change allows DbParams to add arbitrary connection parameters via
environment variables, as opposed to the previous hard-coded values.

The PostGIS Docker image requires POSTGRES_USER as a username.  This is
also shorter.  DbParams helpers have been updated to use `user` instead
of `username` for improved consistency.  Note that this a breaking
change and existing DbParams definitions will need to be updated.