Use iter_rows in copy_rows function
The original implementation of copy_rows iterated over data chunks and
called executemany for each one.  This resulted in excessive log
messages and a resetting of the processed rows counter.  Using iter_rows
to return a row_generator allows all rows to be inserted with a single
call to executemany.