Consolidate integration tests for select methods
Select methods such as get_rows, fetchone call iter_rows with various
parameters, row factories and transforms.  This commit consolidates the
tests so that these are confirmed to work only for iter_rows.
A parameterized test is added to confirm that they are passed to
iter_rows correctly by all the other functions.