Set LD_LIBRARY_PATH via source command
The previous method of setting LD_LIBRARY_PATH was complicated as it
required writing an executable script then finding a directory on the
Python path to symlink it into.  This commit replaces that by writing
a script that can be executed via the "source" command.  This method is
simpler because it requires a single file that can be located in the
installation directory.

This method also be run in a single command, which is always the same.
All installation logging output is sent to <stderr>, except for the name
of the script, which goes to <stdout>.  Using shell command substitution
allows this output to be passed to the "source" command.

This is a breaking change.