Raise warning on DbHelper init without driver
Current behaviour when a DbHelper class is initialised but the required Python
database library e.g. psycopg2 cannt be imported is to print() a message.
This commit changes to warnings.warn().  This keeps the stdout clean and
allows user to suppress the message if desired.

It is necessary to move the setting of self.required_params to before
the library import so that the required params are available to the
DbParams module for validation.