Raise bad dbtype error from None
This commit replaces the previous way of simplifying the commit message
by using the `raise ... from None` syntax.  This achieves the desired
outcome of not including lower messages from the stack in the error
output.  The previous attempt reduced the number of commands within the
reported list for each error in the chain.

See https://www.python.org/dev/peps/pep-0409/