Ensure fetch functions commit current transaction
Connections to the database automatically begin a transaction.
Unless it is closed via a commit() it will be reused in subsequent
transactions.  This causes side effects e.g. the result of the
PostgreSQL now() function remains the same.

This commit updates iter_chunks to commit once the iterator is
exhausted.  fetchone and fetchmany are modfied to commit before
returning as they do not exhaust the iterator.